requirements.txt
numpy
scipy
matplotlib
pandas
jupyter
sklearn
flask
pillow
opencv-python